NITA/requirements.txt
junos-eznc==2.2.1
robotbackgroundlogger==1.2
robotframework-seleniumlibrary==3.0.0a2  
selenium==3.5.0
robotframework==3.0.4
grpcio==1.18.0
protobuf==3.6.1
cryptography==2.7
jxmlease==1.0.1
websocket-client==0.56.0
autopep8==1.4.4    
docutils==0.14
Jinja2==2.10.1       
lazy-object-proxy==1.4.1
lxml==4.3.4
netaddr==0.7.19
PyNaCl==1.3.0
paramiko==2.6.0
pep8==1.7.1
pexpect==4.7.0
pyOpenSSL==19.0.0
pysnmp==4.4.9
PyYAML==5.1.1
robotframework-debuglibrary==1.1.4
scp==0.13.2
Sphinx==2.1.2
sphinx-bootstrap-theme==0.6.5
robotframework-extendedrequestslibrary==0.5.5
boto3==1.9.185
influxdb==5.2.2
troposphere==2.4.9
numpy==1.16.4
pandas==0.24.2
plotly==3.10.0
graphviz==0.11.1
SQLAlchemy==1.3.5
robotframework-run-keyword-async==1.0.8
ntplib==0.3.4
ldap3==2.6
humanfriendly==4.18
python-gitlab==1.9.0
google==2.0.2
pymysql==0.9.3
scapy==2.4.2
ruamel.yaml==0.15.89
pyroute2==0.5.6
pysftp==0.2.9
jsonpath-rw==1.4.0
certifi==2018.11.29
psycopg2==2.8.3
psutil==5.6.3
backports.ssl_match_hostname==3.5.0.1
backports.ssl==0.0.9
socketIO-client==0.7.2 # UR ID : TOBY-4740
robotframework-pabot==1.2.1
